[Feat] Add single session test and move private test to test_request.py
