+++ tests/test_request.py
@@ -11,9 +11,22 @@ def test_method():
     @request("GET", "/{test_path}")
     async def test_request(
         session: Session,
+        test_path: Annotated[str, Path] = "TEST_PATH",
         parameter: Annotated[str, Query] = "TEST_QUERY",
         header: Annotated[str, Header] = "TEST_HEADER",
     ) -> aiohttp.ClientResponse:
+        pass
+
+    return test_request
+
+
+@pytest.fixture
+def test_method_for_private_parameter():
+    @Session.single_session("https://test_base_url")
+    @request("GET", "/")
+    @Header.default_header("private_header", "__PRIVATE_HEADER__")
+    @Query.default_query("private_query", "__PRIVATE_QUERY__")
+    async def test_request(session: Session) -> None:
         pass
 
     return test_request
@@ -33,6 +46,9 @@ def test_copy_and_equal(test_method):
 def test_fill_parameter(test_method):
     new_method = test_method.copy()
 
+    assert "header" in new_method.header_parameter.keys()
+    assert "parameter" in new_method.query_parameter.keys()
+
     assert "header" not in new_method.headers.keys()
     assert "parameter" not in new_method.params.keys()
 
@@ -44,3 +60,29 @@ def test_fill_parameter(test_method):
     assert "parameter" in new_method.params.keys()
     assert new_method.headers["header"] == "TEST_HEADER"
     assert new_method.params["parameter"] == "TEST_QUERY"
+
+
+def test_formatted_path(test_method):
+    new_method = test_method.copy()
+
+    assert "test_path" in new_method.path_parameter.keys()
+
+    bound_argument = test_method._signature.bind(test_method.session)
+    bound_argument.apply_defaults()
+
+    formatted_path = new_method._get_request_path(bound_argument)
+    assert formatted_path == "/TEST_PATH"
+
+
+def test_private_component(test_method_for_private_parameter):
+    assert "private_header" in test_method_for_private_parameter.headers
+    assert "private_query" in test_method_for_private_parameter.params
+
+    assert (
+        test_method_for_private_parameter.headers.get("private_header")
+        == "__PRIVATE_HEADER__"
+    )
+    assert (
+        test_method_for_private_parameter.params.get("private_query")
+        == "__PRIVATE_QUERY__"
+    )
